+++ test/test-proj/App.js
@@ -1,8 +1,14 @@
 "use strict";
 var DataSource = require("./DataSource");
-var HelperUtil = require("./HelperUtil");
+var HelperUtil = require("./helpers/HelperUtil");
+var WidgetUi = require("./WidgetUi");
 function main() {
-    console.log(HelperUtil.App.name + "@" + HelperUtil.App.version);
+    HelperUtil.App = {
+        name: "test-proj",
+        version: "0.1.0",
+        description: "parser/bundler test project",
+    };
+    console.log(WidgetUi.createWidget());
     for (var collName in DataSource.collectionCache) {
         console.log(collName + ": " + DataSource.collectionCache[collName]);
     }