+++ test/test-proj/App.ts
@@ -1,8 +1,15 @@
 ﻿import DataSource = require("./DataSource");
-import HelperUtil = require("./HelperUtil");
+import HelperUtil = require("./helpers/HelperUtil");
+import WidgetUi = require("./WidgetUi");
 
 function main() {
-    console.log(HelperUtil.App.name + "@" + HelperUtil.App.version);
+    HelperUtil.App = {
+        name: "test-proj",
+        version: "0.1.0",
+        description: "parser/bundler test project",
+    };
+
+    console.log(WidgetUi.createWidget());
     for (var collName in DataSource.collectionCache) {
         console.log(collName + ": " + DataSource.collectionCache[collName]);
     }