+++ test/test-proj/WidgetUi.ts
@@ -1,0 +1,13 @@
+﻿import DataSource = require("./DataSource");
+import HelperUtil = require("./helpers/HelperUtil");
+
+module WidgetUi {
+
+    export function createWidget() {
+        var app = HelperUtil.App;
+        return app.name + "@" + app.version + " - " + app.description;
+    }
+
+}
+
+export = WidgetUi;