+++ test/test-proj/helpers/HelperUtil.js
@@ -1,5 +1,7 @@
 "use strict";
+var WidgetUi = require("../WidgetUi");
 var HelperUtil;
 (function (HelperUtil) {
+    HelperUtil.Ui = WidgetUi;
 })(HelperUtil || (HelperUtil = {}));
 module.exports = HelperUtil;